refactor(pages): trim skills and challenge pages under 180-LOC budget
- skills/page.tsx 195 -> 155: extract <AddSkillForm> to _components/
- challenge/page.tsx 200 -> 107: extract parseChallengeFromSearchParams +
  useMonacoPreload to challenge-page-helpers.ts

All four Phase 6 pages (challenge/edit, habits, skills, challenge) now fit
the 180-LOC page target. Lint clean; tests still green.
